+++ packages/wallet/src/initialization/instances/index.ts
@@ -1,4 +1,5 @@
 export { accountsController } from './accounts-controller/accounts-controller';
+export { addressBookController } from './address-book-controller/address-book-controller';
 export { approvalController } from './approval-controller/approval-controller';
 export { connectivityController } from './connectivity-controller/connectivity-controller';
 export { keyringController } from './keyring-controller/keyring-controller';